// src/app/_types/endpoint/enum.ts
export enum queryKeys {
  site = "site",
  regions = "regions",
  countries = "countries",
  cities = "cities",
  applicationTypes = "applicationTypes",
  products = "products",
  assets = "assets",
  asset = "asset",
  parameters = "parameters",
  modules = "modules",
  module = "module",
  sensors = "sensors",
}